requirements.txt
# Data Process Package
pyRawTools
numpy
pandas
scipy
scikit-learn
umap-learn
polars
combat
# Data Visualization Package
matplotlib